feat: Implement Master Formula for credit calculations and owner bookings view
Backend changes:
- Convert credit fields from DECIMAL to INTEGER (migration + models)
- Add calculate-credit-cost endpoint for accurate preview
- Fix credit flow: total_spent only increments on staff approval
- Add GET /timeshare/bookings endpoint for owner to view all bookings
- Store rejection_reason in booking metadata when staff rejects

Frontend changes:
- Replace hardcoded 60x factor with actual backend Master Formula calculation
- Show credit breakdown with season, room type, and per-night cost
- Create MyBookings page (/owner/bookings) with pending/confirmed/rejected sections
- Add rejection reason display with credit refund notification
- Separate "Gastados" (spent/confirmed) from "Pendientes" (pending approval)
- Add Bookings link to sidebar navigation
- Update navigation after booking to /owner/bookings instead of dashboard
- Improve transaction history with status badges (pending/confirmed/refunded)

Credit system improvements:
- Credits now properly tracked: pending until staff approval
- Refunds work correctly without affecting total_spent
- Visual distinction between active, spent, and refunded credits

+++ backend/src/models/CreditTransaction.ts
@@ -170,11 +170,11 @@ CreditTransaction.init(
       allowNull: false,
     },
     amount: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     balance_after: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     status: {
